rename leveldb package to engine
