+++ src/engine/compact.rs
